Restored absolute path to root directory with -e .
When pip-compile runs it replaces "-e ." which is used to make the project
editable with the absolute path on the filesystyem where pip-compile is
run. That make it impossible to use requirements.txt elsewhere.

Restored "-e ." in requirements.txt so that pip install -r requirements
can be run on a remote system, for example Travis CI.